requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.9.0
setuptools==72.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.44.0
whitenoise==6.7.0